dog_analysis_module/requirements.txt
ultralytics
opencv-python
pillow
firebase-admin
python-dotenv
fastapi
uvicorn
python-multipart
aiofiles
torch
torchvision
numpy
requests